requirements.txt
Django==5.1.7
djangorestframework==3.15.2
psycopg2-binary==2.9.10
python-dateutil==2.8.2
Pillow
django-bootstrap5==23.4
django-cors-headers==4.3.1
